minimalegl: support headless rendering
Add the new environment variable QT_QPA_HEADLESS that when
it is set, minimalegl will no longer fail when it cannot
create a surface, allowing creating a QOpenGLContext when
no window manager is present.

For instance, SwiftShader can be used with minimalegl without
X11 on Linux after this change when QT_QPA_HEADLESS is set.